Move invoice related code to invoices package
