+++ app/grandchallenge/invoices/templates/invoices/partials/challenge_invoice_alert_email.md
